refactor(test): 여러 구현체에 대한 테스트를 분리
